fix: 프젝 구조변경
