--- backend/feedbackresult/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class FeedbackresultConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'feedbackresult'
